Set the default inputs
This allows the default input type of frontends that don't call
retro_set_controller_port_device to be RETRO_DEVICE_AUTO instead of
RETRO_DEVICE_NONE.

The Makefile change is just to allow the use of PSS_STYLE to determine
the slash type, instead of only using backslashes for Win32.  This seems
similar to how other cores do it.